Proof of concept for a direct wrapper around a GList of Splits
This commit adds the following for the guile bindings:
- SplitListNode: a type the same memory layout as a GList node
- gnc-account-get-splits: a convenience function that returns a pointer
  to the SplitListNode that is the first split of the account.

The SplitListNode comes with a number of autogenerated support functions:
- SplitListNode-split-set: assign a new split to the SplitListNode
- SplitListNode-split-get: get the split this SplitListNode refers to
- SplitListNode-next-get: returns a SplitListNode for the next split of the given account
- SplitListNode-prev-get: returns a SplitListNode for the previous split of the given account

There are also setter variants for the last two functions. However I'm not sure
if it's safe to use them. That is it may not be safe to alter the list chaining
pointers directly from within scheme. It's probably better to rely on GList's
convenience functions for this instead.

If this is something useful, it probably needs some more finetuning and
definitely some unit tests.

Note the code uses
typedef struct
{
    Split *split;
    struct SplitListNode *next;
    struct SplitListNode *prev;
} SplitListNode;
static SplitListNode *gnc_account_get_splits (const Account *account)
{
    return (SplitListNode*) xaccAccountGetSplitList (account);
}

rather than

typedef struct splitlist_s SplitListNode;
typedef struct
{
    Split *split;
    SplitListNode *next;
    SplitListNode *prev;
} split_list_node;
static SplitListNode *gnc_account_get_splits (const Account *account)
{
    return (SplitListNode*) xaccAccountGetSplitList (account);
}

The former results in consistent use of the typename SplitListNode in guile
where the latter (while avoiding the need to add 'struct' inside the struct)
would result in a mix of SplitListNode and split-list-node in guile.